Run our "freeze", which turns all pages into static html
This will change our server from a dynamic to static site. This is required
because we are moving off of Heroku and the site is too complex. Moving to
static files makes it easy to serve from anywhere.

I need to add workflows for generating the pages each time we make changes, but
for now this is the initial snapshot.